Refactor tests.
Separate each main functional test into its own file. This greatly
simplifies the imports of most of them.

Also simplify the subscriber tests by removing some partially duplicated
code.
